requirements.txt
Flask==2.1.2
mysql-connector-python==8.0.28
python-dotenv==0.20.0
langchain==0.1.8
langchain-community==0.0.21
langchain-core==0.1.24
langchain-openai==0.0.6
groq==0.4.2
langchain-groq==0.0.1
openai==0.11.0